refactor: create files to handle default tools & intents + tool utils
