+++ libs/naas-abi-core/naas_abi_core/services/agent/intents/default_intents.py
@@ -1,0 +1,176 @@
+from naas_abi_core.services.agent.beta.IntentMapper import (
+    Intent,
+    IntentScope,
+    IntentType,
+)
+
+DEFAULT_INTENTS: list = [
+    Intent(
+        intent_value="what's your name?",
+        intent_type=IntentType.AGENT,
+        intent_target="call_model",
+        intent_scope=IntentScope.DIRECT,
+    ),
+    Intent(
+        intent_value="what do you do?",
+        intent_type=IntentType.AGENT,
+        intent_target="call_model",
+        intent_scope=IntentScope.DIRECT,
+    ),
+    Intent(
+        intent_value="comment tu t'appelles?",
+        intent_type=IntentType.AGENT,
+        intent_target="call_model",
+        intent_scope=IntentScope.DIRECT,
+    ),
+    Intent(
+        intent_value="que fais-tu?",
+        intent_type=IntentType.AGENT,
+        intent_target="call_model",
+        intent_scope=IntentScope.DIRECT,
+    ),
+    Intent(
+        intent_value="Hello",
+        intent_type=IntentType.RAW,
+        intent_target="Hello, what can I do for you?",
+        intent_scope=IntentScope.DIRECT,
+    ),
+    Intent(
+        intent_value="Hi",
+        intent_type=IntentType.RAW,
+        intent_target="Hello, what can I do for you?",
+        intent_scope=IntentScope.DIRECT,
+    ),
+    Intent(
+        intent_value="Hey",
+        intent_type=IntentType.RAW,
+        intent_target="Hello, what can I do for you?",
+        intent_scope=IntentScope.DIRECT,
+    ),
+    Intent(
+        intent_value="Salut",
+        intent_type=IntentType.RAW,
+        intent_target="Bonjour, que puis-je faire pour vous?",
+        intent_scope=IntentScope.DIRECT,
+    ),
+    Intent(
+        intent_value="Bonjour",
+        intent_type=IntentType.RAW,
+        intent_target="Bonjour, que puis-je faire pour vous?",
+        intent_scope=IntentScope.DIRECT,
+    ),
+    Intent(
+        intent_value="Coucou",
+        intent_type=IntentType.RAW,
+        intent_target="Bonjour, que puis-je faire pour vous?",
+        intent_scope=IntentScope.DIRECT,
+    ),
+    Intent(
+        intent_value="Hi there",
+        intent_type=IntentType.RAW,
+        intent_target="Hello, what can I do for you?",
+        intent_scope=IntentScope.DIRECT,
+    ),
+    Intent(
+        intent_value="Hello there",
+        intent_type=IntentType.RAW,
+        intent_target="Hello, what can I do for you?",
+        intent_scope=IntentScope.DIRECT,
+    ),
+    Intent(
+        intent_value="Hello, how are you?",
+        intent_type=IntentType.RAW,
+        intent_target="Hello, I am doing well thank you, how can I help you today?",
+        intent_scope=IntentScope.DIRECT,
+    ),
+    Intent(
+        intent_value="Hi, how are you?",
+        intent_type=IntentType.RAW,
+        intent_target="Hello, I am doing well thank you, how can I help you today?",
+        intent_scope=IntentScope.DIRECT,
+    ),
+    Intent(
+        intent_value="Bonjour, comment vas-tu?",
+        intent_type=IntentType.RAW,
+        intent_target="Bonjour, je vais bien merci, comment puis-je vous aider aujourd'hui?",
+        intent_scope=IntentScope.DIRECT,
+    ),
+    Intent(
+        intent_value="Salut, ça va?",
+        intent_type=IntentType.RAW,
+        intent_target="Bonjour, je vais bien merci, comment puis-je vous aider aujourd'hui?",
+        intent_scope=IntentScope.DIRECT,
+    ),
+    Intent(
+        intent_value="Thank you",
+        intent_type=IntentType.RAW,
+        intent_target="You're welcome, can I help you with anything else?",
+        intent_scope=IntentScope.DIRECT,
+    ),
+    Intent(
+        intent_value="Thank you very much",
+        intent_type=IntentType.RAW,
+        intent_target="You're welcome, can I help you with anything else?",
+        intent_scope=IntentScope.DIRECT,
+    ),
+    Intent(
+        intent_value="Thank you so much",
+        intent_type=IntentType.RAW,
+        intent_target="You're welcome, can I help you with anything else?",
+        intent_scope=IntentScope.DIRECT,
+    ),
+    Intent(
+        intent_value="Merci",
+        intent_type=IntentType.RAW,
+        intent_target="Je vous en prie, puis-je vous aider avec autre chose?",
+        intent_scope=IntentScope.DIRECT,
+    ),
+    Intent(
+        intent_value="Merci beaucoup",
+        intent_type=IntentType.RAW,
+        intent_target="Je vous en prie, puis-je vous aider avec autre chose?",
+        intent_scope=IntentScope.DIRECT,
+    ),
+    Intent(
+        intent_value="Merci bien",
+        intent_type=IntentType.RAW,
+        intent_target="Je vous en prie, puis-je vous aider avec autre chose?",
+        intent_scope=IntentScope.DIRECT,
+    ),
+    Intent(
+        intent_value="List tools available",
+        intent_type=IntentType.TOOL,
+        intent_target="list_tools_available",
+        intent_scope=IntentScope.DIRECT,
+    ),
+    Intent(
+        intent_value="What are the tools available?",
+        intent_type=IntentType.TOOL,
+        intent_target="list_tools_available",
+        intent_scope=IntentScope.DIRECT,
+    ),
+    Intent(
+        intent_value="List sub-agents available",
+        intent_type=IntentType.TOOL,
+        intent_target="list_subagents_available",
+        intent_scope=IntentScope.DIRECT,
+    ),
+    Intent(
+        intent_value="What are the sub-agents available?",
+        intent_type=IntentType.TOOL,
+        intent_target="list_subagents_available",
+        intent_scope=IntentScope.DIRECT,
+    ),
+    Intent(
+        intent_value="List intents",
+        intent_type=IntentType.TOOL,
+        intent_target="list_intents_available",
+        intent_scope=IntentScope.DIRECT,
+    ),
+    Intent(
+        intent_value="What are your intents?",
+        intent_type=IntentType.TOOL,
+        intent_target="list_intents_available",
+        intent_scope=IntentScope.DIRECT,
+    ),
+]
